add etag to get, make de module public
